services/sentiment/requirements.txt
finnhub-python
transformers
torch
psycopg2-binary
sqlalchemy
python-dotenv
